Moved options outside the primary folder.
