+++ src/ant.ts
@@ -7,13 +7,13 @@ import { ApiMultipleResultQueryManager } from './api/query/api-multiple-result-q
 import { ApiQueryConfig } from './api/config/api-query-config';
 import { ApiQueryManager } from './api/query/api-query-manager';
 import { ApiSingleResultQueryManager } from './api/query/api-single-result-query-manager';
-import { CacheMode } from './persistence/primary/options/cache-mode';
+import { CacheMode } from './persistence/options/cache-mode';
 import { Entity } from './model/entity';
 import { KeyGenParams } from './model/key-gen-params';
 import { Model } from './model/model';
-import { PersistencyDeleteOptions } from './persistence/primary/options/persistency-delete-options';
-import { PersistencySearchOptions } from './persistence/primary/options/persistency-search-options';
-import { PersistencyUpdateOptions } from './persistence/primary/options/persistency-update-options';
+import { PersistencyDeleteOptions } from './persistence/options/persistency-delete-options';
+import { PersistencySearchOptions } from './persistence/options/persistency-search-options';
+import { PersistencyUpdateOptions } from './persistence/options/persistency-update-options';
 
 export {
   AntManager,
