+++ src/persistence/primary/ant-primary-model-manager.ts
@@ -3,8 +3,8 @@ import { AntPrimaryEntityManager } from './ant-primary-entity-manager';
 import { DeleteEntitiesCachedScriptSet } from './script/delete-entities-cached-script-set';
 import { Entity } from '../../model/entity';
 import { Model } from '../../model/model';
-import { PersistencyDeleteOptions } from './options/persistency-delete-options';
-import { PersistencyUpdateOptions } from './options/persistency-update-options';
+import { PersistencyDeleteOptions } from '../options/persistency-delete-options';
+import { PersistencyUpdateOptions } from '../options/persistency-update-options';
 import { PrimaryModelManager } from './primary-model-manager';
 import { PrimaryQueryManager } from './query/primary-query-manager';
 import { RedisCachedScript } from './script/redis-cached-script';
