+++ src/persistence/primary/query/ant-multiple-result-primary-query-manager.ts
@@ -3,7 +3,7 @@ import { SEPARATOR_STRING, VOID_RESULT_STRING } from '../lua-constants';
 import { AntPrimaryQueryManager } from './ant-primary-query-manager';
 import { Entity } from '../../../model/entity';
 import { MultipleResultPrimaryQueryManager } from './multiple-result-primary-query-manager';
-import { PersistencySearchOptions } from '../options/persistency-search-options';
+import { PersistencySearchOptions } from '../../options/persistency-search-options';
 
 export class AntMultipleResultPrimaryQueryManager<TEntity extends Entity>
   extends AntPrimaryQueryManager<TEntity, number[] | string[]>
