+++ src/persistence/primary/query/ant-single-result-primary-query-manager.ts
@@ -1,7 +1,7 @@
 import * as _ from 'lodash';
 import { AntPrimaryQueryManager } from './ant-primary-query-manager';
 import { Entity } from '../../../model/entity';
-import { PersistencySearchOptions } from '../options/persistency-search-options';
+import { PersistencySearchOptions } from '../../options/persistency-search-options';
 import { SingleResultPrimaryQueryManager } from './single-result-primary-query-manager';
 import { VOID_RESULT_STRING } from '../lua-constants';
