+++ src/persistence/scheduler/scheduler-model-manager.ts
@@ -1,8 +1,8 @@
 import { MultipleQueryResult, SingleQueryResult, TMQuery, TQuery } from '../primary/query/query-types';
 import { Entity } from '../../model/entity';
 import { MultipleResultPrimaryQueryManager } from '../primary/query/multiple-result-primary-query-manager';
-import { PersistencyDeleteOptions } from '../primary/options/persistency-delete-options';
-import { PersistencySearchOptions } from '../primary/options/persistency-search-options';
+import { PersistencyDeleteOptions } from '../options/persistency-delete-options';
+import { PersistencySearchOptions } from '../options/persistency-search-options';
 import { RedisMiddleware } from '../primary/redis-middleware';
 import { SingleResultPrimaryQueryManager } from '../primary/query/single-result-primary-query-manager';
