+++ src/test/primary/query/multiple-result-query-manager-test.ts
@@ -1,7 +1,7 @@
 import { NamedEntity, NamesStartingByLetter } from './names-starting-by-letter';
 import { NamedEntityAlternative, NamesStartingByLetterAlternative } from './names-starting-by-letter-alternative';
-import { AntJsDeleteOptions } from '../../../persistence/primary/options/antjs-delete-options';
-import { AntJsSearchOptions } from '../../../persistence/primary/options/antjs-search-options';
+import { AntJsDeleteOptions } from '../../../persistence/options/antjs-delete-options';
+import { AntJsSearchOptions } from '../../../persistence/options/antjs-search-options';
 import { AntModel } from '../../../model/ant-model';
 import { AntPrimaryModelManager } from '../../../persistence/primary/ant-primary-model-manager';
 import { Model } from '../../../model/model';
